requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
audio_recorder_streamlit
gTTS